eval/detllm/requirements.txt
detllm>=0.1.1
pyyaml>=6.0
